requirements.txt
discord.py == 1.7.3
python-dotenv == 0.20.0
#gspread~3.7.0
#pytz~=2021.1